+++ src/apps/ease/ethereum/ease.rca.token-fetcher.ts
@@ -4,6 +4,7 @@ import _ from 'lodash';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
+import { getAppImg } from '~app-toolkit/helpers/presentation/image.present';
 import { YEARN_DEFINITION } from '~apps/yearn/yearn.definition';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { AppTokenPosition } from '~position/position.interface';
@@ -29,7 +30,7 @@ export class EthereumEaseRcaTokenFetcher implements PositionFetcher<AppTokenPosi
   constructor(
     @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
     @Inject(EaseContractFactory) private readonly easeContractFactory: EaseContractFactory,
-  ) {}
+  ) { }
 
   async getPositions() {
     const endpoint = 'https://app.ease.org/api/v1/vaults';
@@ -60,7 +61,8 @@ export class EthereumEaseRcaTokenFetcher implements PositionFetcher<AppTokenPosi
           .balanceOf(address)
           .then(v => Number(v) / 10 ** underlyingToken.decimals),
       resolvePricePerShare: () => 1,
-      resolveApy: async ({ vaultAddress }) => (await (rcaAddressToDetails[vaultAddress]?.token['apy'] ?? 0)) / 100,
+      resolveApy: async ({ vaultAddress }) => (await (rcaAddressToDetails[vaultAddress]?.token['apy'] ?? 0)),
+      resolveImages: () => [getAppImg(appId)],
     });
   }
 }